+++ packages/hub/src/app/ai/WorkflowViewer/index.tsx
@@ -9,7 +9,7 @@ import { useAvailableHeight } from "@/hooks/useAvailableHeight";
 import { LogsView } from "../LogsView";
 import { SquigglePlaygroundForWorkflow } from "../SquigglePlaygroundForWorkflow";
 import { Header } from "./Header";
-import { WorkflowSteps } from "./WorkflowActions";
+import { WorkflowSteps } from "./WorkflowSteps";
 
 type WorkflowViewerProps<
   T extends ClientWorkflow["status"] = ClientWorkflow["status"],
@@ -59,7 +59,7 @@ const FinishedWorkflowViewer: FC<WorkflowViewerProps<"finished">> = ({
             <div className="flex gap-2">
               <StyledTab.List>
                 <StyledTab name="Playground" />
-                <StyledTab name="Actions" />
+                <StyledTab name="Steps" />
                 <StyledTab name="Logs" />
               </StyledTab.List>
             </div>
